pasando la primera funcion a entero

+++ Taller 1/src/main/scala/Recursion/package.scala
@@ -1,13 +1,13 @@
 package object Recursion {
   //Calcular el minimo comun divisor mediante el TFA
-  def mcdTFA(ln: List[Int], lm: List[Int], primos: List[Int]): Double = {
+  def mcdTFA(ln: List[Int], lm: List[Int], primos: List[Int]): Int = {
     def cmpEXPO(ln: List[Int],lm: List[Int]):Int = if (ln.head>lm.head) lm.head else ln.head
     def pwrPrime(primes: List[Int], ln: List[Int], lm: List[Int]): Double = scala.math.pow(primes.head,cmpEXPO(ln, lm))
     def pwrs(primes: List[Int], ln: List[Int], lm: List[Int], result: Double):Double= {
       if (primes.isEmpty) result
       else pwrs(primes.tail,ln.tail,lm.tail,result*pwrPrime(primes, ln, lm))
     }
-    pwrs(primos, ln, lm,1)
+    pwrs(primos, ln, lm,1).toInt
   }
   //Calcular el minimo comun divisor y coneficientes de berzuit mediante el algoritmo de Euclides
   def mcdEB(n: Int, m: Int): Int = {
